record.NewJSONRecordFromDecoder added back
The issue with NewJSONRecordFromReader is that we're creating a new
decoder for each record, and JSON decoders consume more data than needed
from the underlying reader, keeping the trailing data in a buffer.
This means that the first created record would consume a large chunk of
the data, causing EOF errors on small inputs.